+++ src/main/resources/templates/domain/mypage/cards.html
@@ -27,6 +27,11 @@
     <div class="max-w-4xl">
       <h1 class="text-2xl text-gray-800 mb-8">카드 관리</h1>
 
+      <div th:if="${errorMsg}" class="mb-6 p-4 bg-red-50 border border-red-200 text-red-700 rounded-xl flex items-center gap-2">
+        <i data-lucide="alert-circle" class="w-5 h-5"></i>
+        <span th:text="${errorMsg}"></span>
+      </div>
+
       <div id="cardNotIssued" th:if="${!cardIssued}">
         <div class="bg-gradient-to-r from-[#FCE8E3] to-[#FFE4E8] rounded-3xl shadow-sm p-8 mb-6">
           <div class="flex items-center justify-between">
@@ -34,10 +39,10 @@
               <h2 class="text-xl text-gray-800 mb-2">새 카드 발급</h2>
               <p class="text-gray-600">모담 체크카드를 발급받으세요</p>
             </div>
-            <a th:href="@{/mypage/card/step1}" class="px-8 py-3 bg-[#F5A5A5] text-white rounded-2xl hover:bg-[#F28B8B] transition-colors shadow-md flex items-center gap-2">
+            <button onclick="handleMenuClick(event, '/mypage/card/step1')" class="px-8 py-3 bg-[#F5A5A5] text-white rounded-2xl hover:bg-[#F28B8B] transition-colors shadow-md flex items-center gap-2">
               <i data-lucide="credit-card" class="w-5 h-5"></i>
               <span>카드 발급하기</span>
-            </a>
+            </button>
           </div>
         </div>
         <div class="bg-white rounded-3xl shadow-sm p-8">
@@ -47,7 +52,7 @@
               <i data-lucide="credit-card" class="w-10 h-10 text-gray-400"></i>
             </div>
             <p class="text-gray-500 mb-4">등록된 카드가 없습니다</p>
-            <a th:href="@{/mypage/card/step1}" class="px-6 py-2 bg-[#FCE8E3] text-gray-800 rounded-xl hover:bg-[#FBD5CE] transition-colors">첫 카드 발급하기</a>
+            <button onclick="handleMenuClick(event, '/mypage/card/step1')" class="px-6 py-2 bg-[#FCE8E3] text-gray-800 rounded-xl hover:bg-[#FBD5CE] transition-colors">첫 카드 발급하기</button>
           </div>
         </div>
       </div>
@@ -149,7 +154,10 @@
               <button onclick="handleReportLost()" class="py-3 bg-red-50 text-red-700 rounded-xl hover:bg-red-100 transition-colors">분실 신고로 전환</button>
             </div>
             <div id="actionsLost" class="hidden">
-              <a th:href="@{/mypage/card/step1}" class="w-full block text-center py-3 bg-[#F5A5A5] text-white rounded-xl hover:bg-[#F28B8B] transition-colors">재발급 신청</a>
+              <button onclick="handleMenuClick(event, '/mypage/card/step1')" class="w-full block text-center py-3 bg-[#F5A5A5] text-white rounded-xl hover:bg-[#F28B8B] transition-colors">재발급 신청</button>
+            </div>
+            <div id="actionsInactive" class="hidden">
+              <button onclick="alert('해지된 계좌의 카드는 관리할 수 없습니다.')" class="w-full py-3 bg-gray-200 text-gray-500 rounded-xl cursor-not-allowed font-medium">카드 기능 사용 불가 (계좌 해지됨)</button>
             </div>
           </div>
         </div>
@@ -168,15 +176,15 @@
               <p class="text-gray-800 font-medium group-hover:text-[#F28B8B] transition-colors">최근 결제 내역 확인</p>
             </a>
 
-            <a th:href="@{/mypage/notifications}" class="py-4 bg-[#FAFAFA] text-gray-800 rounded-2xl hover:bg-[#FCE8E3] transition-colors text-left px-5 block group">
+            <button onclick="handleMenuClick(event, '/mypage/notifications')" class="py-4 bg-[#FAFAFA] text-gray-800 rounded-2xl hover:bg-[#FCE8E3] transition-colors text-left px-5 block group w-full">
               <p class="text-sm text-[#F5A5A5] mb-1 flex items-center gap-1"><i data-lucide="bell-ring" class="w-4 h-4"></i> 결제 알림 설정</p>
               <p class="text-gray-800 font-medium group-hover:text-[#F28B8B] transition-colors">푸시 및 알림 관리</p>
-            </a>
+            </button>
 
-            <a th:href="@{/mypage/card/step1}" class="py-4 bg-[#FAFAFA] text-gray-800 rounded-2xl hover:bg-[#FCE8E3] transition-colors text-left px-5 block group">
+            <button onclick="handleMenuClick(event, '/mypage/card/step1')" class="py-4 bg-[#FAFAFA] text-gray-800 rounded-2xl hover:bg-[#FCE8E3] transition-colors text-left px-5 block group w-full">
               <p class="text-sm text-[#F5A5A5] mb-1 flex items-center gap-1"><i data-lucide="refresh-cw" class="w-4 h-4"></i> 카드 재발급</p>
               <p class="text-gray-800 font-medium group-hover:text-[#F28B8B] transition-colors">새로운 카드로 교체</p>
-            </a>
+            </button>
 
           </div>
         </div>
@@ -247,8 +255,8 @@
 <script th:inline="javascript">
   const serverCardDesign = /*[[${cardDesign}]]*/ 'pink';
   const serverCardType   = /*[[${cardType}]]*/ 'domestic';
-  // 🔥 백엔드에서 주입한 동적 CVV 변수 받기
   const serverCvv        = /*[[${cardCvv != null ? cardCvv : '123'}]]*/ '123';
+  const serverAccountClosed = /*[[${isAccountClosed}]]*/ false;
   let cardStatus         = /*[[${cardStatus}]]*/ 'ACTIVE';
 
   const cardColors = {
@@ -298,12 +306,14 @@
     cardInner.classList.toggle('flipped');
   }
 
+  // 🔥 [수정됨] INACTIVE(해지됨) 배지 및 색상 로직 추가
   function updateCardStatus(status) {
     cardStatus = status;
     const banners = {
       ACTIVE:  { bg:'bg-green-50', border:'border-green-200', icon:'✅', title:'모든 기능 사용 가능', desc:'정상 작동 중입니다', badge:'정상', badgeCls:'bg-green-50 text-green-600', dot:'bg-green-500' },
       STOPPED: { bg:'bg-orange-50', border:'border-orange-200', icon:'⏸️', title:'사용 일시 중지', desc:'언제든 재개할 수 있어요', badge:'일시 정지', badgeCls:'bg-orange-50 text-orange-600', dot:'bg-orange-500' },
-      LOST:    { bg:'bg-red-50', border:'border-red-200', icon:'🚨', title:'사용 불가', desc:'재발급 신청이 필요해요', badge:'분실 신고됨', badgeCls:'bg-red-50 text-red-600', dot:'bg-red-500' }
+      LOST:    { bg:'bg-red-50', border:'border-red-200', icon:'🚨', title:'사용 불가', desc:'재발급 신청이 필요해요', badge:'분실 신고됨', badgeCls:'bg-red-50 text-red-600', dot:'bg-red-500' },
+      INACTIVE:{ bg:'bg-gray-100', border:'border-gray-200', icon:'🚫', title:'사용 불가', desc:'계좌가 해지되어 사용할 수 없습니다', badge:'해지됨', badgeCls:'bg-gray-200 text-gray-500', dot:'bg-gray-400' }
     };
     const s = banners[status] || banners.ACTIVE;
     const banner = document.getElementById('cardStatusBanner');
@@ -326,9 +336,10 @@
       document.getElementById('cardOverlayText').textContent = status === 'STOPPED' ? '일시 정지' : '사용 불가';
     }
 
-    document.getElementById('actionsActive').classList.toggle('hidden', status !== 'ACTIVE');
-    document.getElementById('actionsStopped').classList.toggle('hidden', status !== 'STOPPED');
-    document.getElementById('actionsLost').classList.toggle('hidden', status !== 'LOST');
+    document.getElementById('actionsActive')?.classList.toggle('hidden', status !== 'ACTIVE');
+    document.getElementById('actionsStopped')?.classList.toggle('hidden', status !== 'STOPPED');
+    document.getElementById('actionsLost')?.classList.toggle('hidden', status !== 'LOST');
+    document.getElementById('actionsInactive')?.classList.toggle('hidden', status !== 'INACTIVE');
   }
 
   function toggleCardNumber() {
@@ -348,7 +359,6 @@
     lucide.createIcons();
   }
 
-  // 🔥 [수정] 하드코딩 123 제거 후 동적 CVV 토글 적용
   function toggleCVV() {
     const el = document.getElementById('cvvDisplay');
     const icon = document.getElementById('eyeCvv');
@@ -356,6 +366,16 @@
     el.textContent = showing ? '***' : serverCvv;
     icon.setAttribute('data-lucide', showing ? 'eye' : 'eye-off');
     lucide.createIcons();
+  }
+
+  // 🔥 [추가됨] 계좌 해지 시 메뉴 클릭(발급, 알림 설정 등) 차단 로직
+  function handleMenuClick(event, url) {
+    if (serverAccountClosed) {
+        event.preventDefault();
+        alert('해지된 계좌에서는 해당 기능을 이용할 수 없습니다.');
+        return;
+    }
+    window.location.href = url;
   }
 
   function handleStopCard() {
