+++ src/main/resources/templates/domain/mypage/close-account.html
@@ -57,7 +57,7 @@
                                 <span th:if="${accountClosureStatus == 'closed'}" class="px-3 py-1 bg-gray-100 text-gray-700 text-xs rounded-lg">해지됨</span>
                             </div>
                             <div class="space-y-1 text-sm text-gray-600">
-                                <div>계좌번호: <span class="font-mono">1002-9876-5432</span></div>
+                                <div>계좌번호: <span class="font-mono" th:text="${accountNumber != null and accountNumber != '' ? accountNumber : '1002-9876-5432'}">1002-9876-5432</span></div>
                                 <div>잔액: <span class="text-xl font-bold text-gray-900">₩<span th:text="${accountBalance != null ? #numbers.formatInteger(accountBalance, 0, 'COMMA') : '0'}" id="closeAccountBalance">0</span></span></div>
                                 <div>개설일: 2026.01.15</div>
                                 <div>파트너: <span th:text="${partnerName != null ? partnerName : '대기 중'}" id="closePartnerName">대기 중</span></div>
@@ -227,7 +227,6 @@
       lucide.createIcons();
     });
 
-    // 🔥 [수정] 해지 완료(closed) 상태일 때 버튼 및 배너를 회색으로 비활성화하는 로직 추가
     function initClosureSection() {
       const banner = document.getElementById('closureStatusBanner');
       const btn    = document.getElementById('closureActionBtn');
@@ -241,9 +240,13 @@
         document.getElementById('closureStatusDesc').className = 'text-sm leading-relaxed text-gray-600';
         document.getElementById('closureStatusDesc').textContent  = '이 계좌는 이미 성공적으로 해지 및 정산이 완료되었습니다.';
 
-        btn.className = 'w-full py-4 rounded-xl transition-colors font-medium text-lg bg-gray-300 text-white cursor-not-allowed';
-        btn.textContent = '이미 해지된 계좌입니다';
-        btn.disabled = true;
+        btn.className = 'w-full py-4 rounded-xl transition-colors font-medium text-lg bg-[#F5A5A5] hover:bg-[#F28B8B] text-white shadow-md flex justify-center items-center gap-2';
+        btn.innerHTML = '새로운 커플 통장 만들기 <i data-lucide="arrow-right" class="w-5 h-5"></i>';
+        btn.disabled = false;
+
+        btn.onclick = function() {
+            window.location.href = '/group-account/new';
+        };
 
       } else if (closureStatus === 'pending') {
         banner.classList.remove('hidden');
@@ -258,6 +261,7 @@
           btn.className = 'w-full py-4 rounded-xl transition-colors font-medium text-lg bg-gray-500 hover:bg-gray-600 text-white';
           btn.textContent = '해지 요청 취소하기';
           btn.disabled = false;
+          btn.onclick = handleClosureAction;
         } else {
           banner.className = 'rounded-xl p-6 border-2 bg-orange-50 border-orange-200';
           document.getElementById('closureStatusIcon').textContent  = '⚠️';
@@ -269,17 +273,18 @@
           btn.className = 'w-full py-4 rounded-xl transition-colors font-medium text-lg bg-orange-500 hover:bg-orange-600 text-white';
           btn.textContent = '해지 동의하기';
           btn.disabled = false;
+          btn.onclick = handleClosureAction;
         }
       } else {
         banner.classList.add('hidden');
         btn.className = 'w-full py-4 rounded-xl transition-colors font-medium text-lg bg-red-500 hover:bg-red-600 text-white';
         btn.textContent = '커플 통장 해지 요청하기';
         btn.disabled = false;
+        btn.onclick = handleClosureAction;
       }
     }
 
     function handleClosureAction() {
-      // 해지 완료 상태면 클릭 무시
       if (closureStatus === 'closed') return;
 
       if (closureStatus === 'pending' && closureRequestedBy === userName) {
@@ -367,6 +372,5 @@
       document.getElementById('refundPartnerContrib3').textContent = '기여도 ' + prC.toFixed(1) + '% 기준';
     }
 </script>
-<th:block th:replace="~{common/layout/header :: scripts}"></th:block>
 </body>
 </html>
