+++ src/main/resources/templates/domain/mypage/profile.html
@@ -95,6 +95,28 @@
                                    class="w-full px-4 py-3 border-2 border-gray-200 rounded-xl focus:border-[#FCE8E3] focus:outline-none transition-colors"
                                    oninput="formatPhone(this)" maxlength="13" required/>
                         </div>
+
+                        <div>
+                            <label class="block text-sm text-gray-600 mb-2">주소</label>
+                            <div class="space-y-3">
+                                <div class="flex gap-3">
+                                    <input type="text" name="zipCode" id="postalCode" th:value="${member != null ? member.zipCode : ''}"
+                                           class="w-32 px-4 py-3 border-2 border-gray-200 rounded-xl focus:border-[#FCE8E3] focus:outline-none transition-colors bg-gray-50 focus:bg-white"
+                                           placeholder="우편번호" readonly/>
+                                    <button type="button" onclick="searchAddress()"
+                                            class="px-4 py-3 bg-gray-100 text-gray-700 rounded-xl hover:bg-gray-200 transition-colors whitespace-nowrap text-sm font-medium">
+                                        주소 검색
+                                    </button>
+                                </div>
+                                <input type="text" name="address" id="address" th:value="${member != null ? member.address : ''}"
+                                       class="w-full px-4 py-3 border-2 border-gray-200 rounded-xl focus:border-[#FCE8E3] focus:outline-none transition-colors bg-gray-50 focus:bg-white"
+                                       placeholder="기본주소" readonly/>
+                                <input type="text" name="addressDetail" id="addressDetail" th:value="${member != null ? member.addressDetail : ''}"
+                                       class="w-full px-4 py-3 border-2 border-gray-200 rounded-xl focus:border-[#FCE8E3] focus:outline-none transition-colors bg-gray-50 focus:bg-white"
+                                       placeholder="상세주소"/>
+                            </div>
+                        </div>
+
                         <button type="submit" class="w-full py-3 bg-[#FCE8E3] text-gray-800 rounded-xl hover:bg-[#FBD5CE] transition-colors mt-4 font-medium">정보 수정 저장하기</button>
                     </div>
                 </form>
@@ -156,6 +178,7 @@
 
 <th:block th:replace="~{common/layout/footer :: footer}"></th:block>
 
+<script src="//t1.daumcdn.net/mapjsapi/bundle/postcode/prod/postcode.v2.js"></script>
 <script>
     function formatPhone(input) {
       const digits = input.value.replace(/[^0-9]/g, '').slice(0, 11);
@@ -164,10 +187,19 @@
       else                         input.value = digits.slice(0, 3) + '-' + digits.slice(3, 7) + '-' + digits.slice(7);
     }
 
+    // 🔥 [추가됨] 주소 검색창 호출 함수
+    function searchAddress() {
+      new daum.Postcode({
+        oncomplete: function(data) {
+          document.getElementById('postalCode').value = data.zonecode;
+          document.getElementById('address').value = data.roadAddress || data.jibunAddress;
+          document.getElementById('addressDetail').focus(); // 상세 주소로 포커스 자동 이동
+        }
+      }).open();
+    }
+
     window.addEventListener('DOMContentLoaded', () => {
         lucide.createIcons();
-
-        // 🔥 [추가됨] 페이지 렌더링 즉시 서버에서 가져온 초기 전화번호에 하이픈 포맷 적용
         const phoneInput = document.getElementById('profilePhone');
         if (phoneInput && phoneInput.value) {
             formatPhone(phoneInput);
